% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-quality_checks.R
\name{add_mad_med_columns}
\alias{add_mad_med_columns}
\title{Add Median and MAD Columns for Indicators}
\usage{
add_mad_med_columns(.data, indicators, group_by = "district")
}
\arguments{
\item{.data}{A \code{cd_data} tibble containing health indicator data.}

\item{indicators}{A character vector specifying the names of the indicator
columns for which the median and MAD should be calculated.}

\item{group_by}{A character string or vector specifying the column(s) to group
by when calculating the median and MAD (default is \code{'district'}).}
}
\value{
A tibble with additional columns for each indicator, named
\verb{\{indicator\}_med} and \verb{\{indicator\}_mad}, containing the calculated median
and MAD, respectively.
}
\description{
\code{add_mad_med_columns} calculates and appends the median and Median Absolute
Deviation (MAD) for specified indicators in a dataset. These statistics are
calculated per group, allowing the user to dynamically group data (e.g., by
\code{district}) and exclude data from the most recent year in the dataset.
}
\details{
\itemize{
\item The median and MAD are calculated for each group specified by \code{group_by}.
\item Only data from years prior to the most recent year in the dataset are considered.
\item Missing values in the calculated statistics are replaced using \code{robust_max()},
ensuring that meaningful fallback values are provided.
}
}
\examples{
\dontrun{
  # Add median and MAD columns for 'indicator1'
  add_mad_med_columns(data, indicators = 'indicator1',
                    group_by = 'district')
}

}
\seealso{
\itemize{
\item \code{\link[=add_outlier5std_column]{add_outlier5std_column()}} for generating outlier flags based on these columns.
\item \code{\link[=robust_max]{robust_max()}} for calculating max value
}
}
