% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_national_coverage.R
\name{analyze_national_coverage}
\alias{analyze_national_coverage}
\title{Analyze National Immunization Coverage Data}
\usage{
analyze_national_coverage(
  .data,
  country_name,
  country_iso,
  indicator = c("bcg", "anc1", "pcv3", "opv1", "opv2", "opv3", "penta2", "pcv1", "pcv2",
    "penta1", "penta3", "measles1", "rota1", "rota2", "instdeliveries", "measles2",
    "ipv1", "ipv2", "undervax", "dropout_penta13", "zerodose", "dropout_measles12",
    "dropout_penta3mcv1"),
  denominator = c("dhis2", "anc1", "penta1")
)
}
\arguments{
\item{.data}{A data frame of type \code{cd_data} containing immunization coverage data.}

\item{country_name}{Character. Name of the country, e.g., "Kenya".}

\item{country_iso}{Character. ISO3 country code, e.g., "KEN".}

\item{indicator}{Character. The immunization indicator to analyze. Must be one
of the supported values such as "bcg", "penta3", etc.
Supported indicators include:
\itemize{
\item General vaccinations: "bcg", "anc1", "pcv3", "opv1", "opv2", "opv3"
\item Dose-based vaccinations: "penta2", "penta1", "penta3", "measles1", "measles2"
\item Additional vaccines: "rota1", "rota2", "ipv1", "ipv2"
\item Other health indicators: "undervax", "dropout_penta13", "zerodose",
"dropout_measles12", "dropout_penta3mcv1"
}}

\item{denominator}{Character. The denominator used for calculating coverage
estimates. Must be one of "dhis2", "anc1", or "penta1".}
}
\value{
A data frame of class \code{cd_national_coverage}, containing year-wise
immunization coverage estimates from DHIS2, WUENIC, and Survey data, along
with confidence intervals if available. The table provides organized coverage
metrics and includes a new \code{cd_national_coverage} class for enhanced functionality.
}
\description{
This function, \code{analyze_national_coverage}, calculates and combines national
immunization coverage data for a specified country, indicator, and denominator.
It integrates internal data sources (\code{survey_data}, \code{wuenic_data})and processes
the provided data source (\code{.data}).
}
\details{
The function performs the following steps:
\enumerate{
\item Validates \code{.data} and verifies required columns.
\item Extracts the appropriate data columns from national, survey, and WUENIC data.
\item Combines all relevant data into a single long-format table, including
estimates and confidence intervals.
\item Creates a new tibble with class \code{cd_national_coverage}, adding metadata
such as the chosen denominator.
}

This function provides a flexible way to analyze immunization coverage by
offering year-wise summaries and estimate categories. The final table format
is ready for further analysis or visualization.
}
\examples{
\dontrun{
analyze_national_coverage(.data = my_data, country_name = "Kenya",
                          country_iso = "KEN", indicator = "penta3",
                          denominator = "penta1")
}

}
