% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{attr_or_abort}
\alias{attr_or_abort}
\title{Retrieve Attribute or Abort}
\usage{
attr_or_abort(.data, attr_name)
}
\arguments{
\item{.data}{An object}

\item{attr_name}{The name of the attribute to retrieve}
}
\value{
The attribute value
}
\description{
Safely retrieves an attribute from an object or aborts with an error.
}
\keyword{internal}
