% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_overall_score.R
\name{calculate_overall_score}
\alias{calculate_overall_score}
\title{Calculate Overall Quality Score}
\usage{
calculate_overall_score(.data)
}
\arguments{
\item{.data}{A data frame of type \code{cd_data}.}
}
\description{
Calculate Overall Quality Score
}
