% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_reporting_rate.R
\name{calculate_reporting_rate}
\alias{calculate_reporting_rate}
\title{\code{calculate_reporting_rate} calculates reporting rate by administrative region
for a specified indicator}
\usage{
calculate_reporting_rate(
  .data,
  indicator,
  admin_level = c("adminlevel_1", "district"),
  threshold = 90
)
}
\arguments{
\item{.data}{A tibble of class \code{cd_data} containing reporting rates.}

\item{indicator}{Character. The indicator to be evaluated, Must be one of
\code{"anc_rr"}, \code{"idelv_rr"}, or \code{"vacc_rr"}}

\item{admin_level}{Character. The administrative level at which to assess reporting rates.
Must be one of \code{"adminlevel_1"} or \code{"district"}.}

\item{threshold}{Numeric. The reporting rate threshold below which administrative units
will be flagged. Defaults to 90.}
}
\value{
A tibble containing administrative units and years where reporting rates for any
indicator group fall below the specified threshold.
}
\description{
\code{calculate_reporting_rate} calculates reporting rate by administrative region
for a specified indicator
}
