% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indicator_threshold_coverage.R
\name{calculate_threshold}
\alias{calculate_threshold}
\title{Calculate Dropout Coverage for Health Indicators Below a Threshold}
\usage{
calculate_threshold(
  .data,
  admin_level = c("adminlevel_1", "district"),
  indicator = c("coverage", "dropout"),
  sbr = 0.02,
  nmr = 0.025,
  pnmr = 0.024,
  anc1survey = 0.98,
  dpt1survey = 0.97,
  survey_year = 2019,
  twin = 0.015,
  preg_loss = 0.03
)
}
\arguments{
\item{.data}{A tibble of class \code{cd_data}.}

\item{admin_level}{The level of analysis.}

\item{indicator}{Character. The specific health indicator to evaluate. Options are:
\itemize{
\item \code{"coverage"}:coverage indicators.
\item \code{"dropout"}: dropout indicators.
}}

\item{sbr}{Numeric. The stillbirth rate. Default is \code{0.02}.}

\item{nmr}{Numeric. Neonatal mortality rate. Default is \code{0.025}.}

\item{pnmr}{Numeric. Post-neonatal mortality rate. Default is \code{0.024}.}

\item{anc1survey}{Numeric. Survey-derived coverage rate for ANC-1 (antenatal care, first visit). Default is \code{0.98}.}

\item{dpt1survey}{Numeric. Survey-derived coverage rate for Penta-1 (DPT1 vaccination). Default is \code{0.97}.}

\item{survey_year}{Integer. The year of Penta-1 survey provided}

\item{twin}{Numeric. Twin birth rate. Default is \code{0.015}.}

\item{preg_loss}{Numeric. Pregnancy loss rate}
}
\value{
A tibble with the selected administrative level and coverage value for regions
that do not meet the below-10\% threshold for the specified indicator and year. If no regions
meet the criteria, a default row is returned with "None" and 0 as values.
}
\description{
This function filters health indicator data to identify the percentage of administrative
regions where the coverage for a specified indicator falls below a 10\% threshold for a given year.
If no regions meet the criteria (i.e., all values are below the threshold), a default output is returned.
}
\examples{
\dontrun{
  # Example usage:
  result <- calculate_threshold(data, filter_year = 2023, indicator = "zerodose", source = "dhis2")
  result
}

}
