% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_average_reporting_rate}
\alias{check_average_reporting_rate}
\title{Check Average Reporting Rate}
\usage{
check_average_reporting_rate(data)
}
\arguments{
\item{data}{A data frame of class \code{cd_data}.}
}
\value{
A data frame with the average reporting rates and a mean reporting rate
for each year.
}
\description{
This function calculates the average reporting rate for each indicator across
all years.
}
\examples{
\dontrun{

  check_average_reporting_rate(data)
}
}
