% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_outlier.R
\name{check_extreme_outlier}
\alias{check_extreme_outlier}
\title{Check for Extreme Outliers}
\usage{
check_extreme_outlier(data)
}
\arguments{
\item{data}{A data frame containing indicator data.}
}
\value{
A data frame summarizing the percentage of districts without extreme
outliers, by year.
}
\description{
This function identifies extreme outliers by calculating the maximum outlier
value for each indicator in each district and year.
}
\examples{
\dontrun{

  check_extreme_outlier(data)
}
}
