% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_missing.R
\name{check_no_missing_district}
\alias{check_no_missing_district}
\title{Check for Missing Values by District}
\usage{
check_no_missing_district(.data)
}
\arguments{
\item{.data}{A data frame containing missing value indicators for districts and
years.}
}
\value{
A data frame summarizing the percentage of districts with no missing
values by year.
}
\description{
This function calculates the percentage of districts with no missing values
for each year.
}
\examples{
\dontrun{

  check_no_missing_district(data)
}
}
