% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_no_missing_value}
\alias{check_no_missing_value}
\title{Check for Missing Values by Year}
\usage{
check_no_missing_value(data)
}
\arguments{
\item{data}{A data frame containing missing value indicators for districts and
years.}
}
\value{
A data frame summarizing the percentage of districts with no missing
values by year.
}
\description{
This function calculates the percentage of districts with no missing values
for each year.
}
\examples{
\dontrun{

  check_no_missing_value(data)
}
}
