% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_missing.R
\name{check_no_missing_year}
\alias{check_no_missing_year}
\title{Check for Missing Values in Each Year}
\usage{
check_no_missing_year(.data)
}
\arguments{
\item{.data}{A data frame containing missing value indicators for multiple years.}
}
\value{
A data frame summarizing the percentage of non-missing values by year.
}
\description{
This function calculates the percentage of non-missing values for all
indicators for each year.
}
\examples{
\dontrun{

  check_no_missing_year(data)
}
}
