% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_outliers}
\alias{check_outliers}
\title{Check for Outliers}
\usage{
check_outliers(data)
}
\arguments{
\item{data}{A data frame containing indicator data for districts and years.}
}
\value{
A data frame summarizing the percentage of non-outliers for each
indicator, by year.
}
\description{
This function checks for outliers based on the Median Absolute Deviation (MAD)
for a range of indicators. It flags outliers if values fall beyond five
standard deviations from the median.
}
\examples{
\dontrun{

  check_outliers(data)
}
}
