% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_reporting.R
\name{check_reporting_rates}
\alias{check_reporting_rates}
\title{Check Reporting Rates by District}
\usage{
check_reporting_rates(data, threshold = 90)
}
\arguments{
\item{data}{A data frame containing reporting rates for multiple districts
and years.}

\item{threshold}{Threshold to calculate the reporting rates}
}
\value{
A data frame with the percentage of districts with reporting rates
above the threshold for each indicator.
}
\description{
This function calculates the percentage of districts reporting rates above a
defined threshold for each by year.
}
\examples{
\dontrun{

  check_reporting_rates(data)
}
}
