% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey_data.R
\name{create_un_estimates}
\alias{create_un_estimates}
\title{Create UN Estimates Tibble}
\usage{
create_un_estimates(.data, country_iso, start_year, end_year)
}
\arguments{
\item{.data}{A data frame read from the UN estimates file.}

\item{country_iso}{Character. ISO3 country code.}

\item{start_year}{Integer. Starting year of data to keep.}

\item{end_year}{Integer. Ending year of data to keep.}
}
\value{
A tibble of class \code{cd_un_estimates}.
}
\description{
Filters, cleans, and renames columns for UN estimate data and assigns a custom
class.
}
