% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiplot.R
\name{equiplot_education}
\alias{equiplot_education}
\title{A Specialized Dot Plot for Maternal Education Analysis}
\usage{
equiplot_education(.data, indicator, x_title = NULL, dot_size = NULL)
}
\arguments{
\item{.data}{A data frame containing the data to be plotted.}

\item{indicator}{A string specifying the indicator to be analyzed (e.g., 'sba').}

\item{x_title}{Optional. A title for the x-axis. Defaults to '\if{html}{\out{<indicator>}} Coverage (\%)'.}
}
\value{
A ggplot object representing the dot plot.
}
\description{
\code{equiplot_education} generates a dot plot comparing coverage across maternal education levels
(no education, primary, and secondary or higher) for a specific indicator across years.
}
\examples{
\dontrun{
  equiplot_education(data, indicator = 'sba')
}

}
