% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_checks_report.R
\name{generate_checks_report}
\alias{generate_checks_report}
\title{Generate and Export Checks Report}
\usage{
generate_checks_report(
  .data,
  output_file,
  output_format = c("html_document", "word_document"),
  threshold = 90
)
}
\arguments{
\item{.data}{A data frame or list object containing the dataset to be analyzed.}

\item{output_file}{A character string specifying the name and path of the output file.}

\item{output_format}{A character vector specifying the output format, either \code{"html_document"} or \code{"word_document"}.
Default is \code{c("html_document", "word_document")}.}

\item{threshold}{A numeric value indicating the threshold for reporting rates in percentage (default is 90).}
}
\value{
The function renders the report to the specified file in the chosen format.
}
\description{
The \code{generate_checks_report} function generates a data quality report based on a specified dataset,
including metrics such as reporting rates, internal consistency, adequate ratios, outliers,
and missing values. The report is rendered using an RMarkdown template and can be output
in either HTML or Word format.
}
\examples{
\dontrun{
  # Generate a report for the dataset 'data' in Word format with a threshold of 85\%
  checks_report(data, output_file = "data_quality_report.docx",
    output_format = "word_document", threshold = 85)
}
}
