% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reports.R
\name{generate_report}
\alias{generate_report}
\title{Generate and Export Checks Report}
\usage{
generate_report(
  cache,
  output_file,
  report_name,
  output_format = c("word_document", "pdf_document", "html_document")
)
}
\arguments{
\item{cache}{A data of class \code{CacheConnection}.}

\item{output_file}{A character string specifying the name and path of the output file.}

\item{report_name}{The name of report to generate}

\item{output_format}{A character vector specifying the output format, either
\code{'html_document'} or \code{'word_document'}.}
}
\value{
The function renders the report to the specified file in the chosen format.
}
\description{
The \code{generate_report} function generates the final report based on a
specified dataset and parameters. The report is rendered using an RMarkdown
template and can be output in either Word, pdf of html formats.
}
\examples{
\dontrun{
  # Generate a report for the dataset 'data' in Word format with a threshold of 85\%
  checks_report(data, output_file = 'data_quality_report.docx',
    output_format = 'word_document', threshold = 85)
}
}
