% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{get_all_indicators}
\alias{get_all_indicators}
\title{Get All Indicators}
\usage{
get_all_indicators()
}
\value{
Character vector of all indicators
}
\description{
Flatten all indicators from all groups
}
