% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{get_country_iso3}
\alias{get_country_iso3}
\title{Get Country ISO3 Code}
\usage{
get_country_iso3(.data)
}
\arguments{
\item{.data}{A \code{cd_data} object}
}
\value{
The ISO3 country code as a character string
}
\description{
Get Country ISO3 Code
}
