% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{get_country_name}
\alias{get_country_name}
\title{Get Country Name}
\usage{
get_country_name(.data)
}
\arguments{
\item{.data}{A \code{cd_data} object}
}
\value{
The country name as a character string
}
\description{
Get Country Name
}
