% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_mapping_get_mapping_data.R
\name{get_country_shapefile}
\alias{get_country_shapefile}
\title{Retrieve Country Shapefile}
\usage{
get_country_shapefile(country_iso, level = c("admin_level_1", "district"))
}
\arguments{
\item{country_iso}{A string specifying the ISO3 code of the country (e.g., 'KEN' for Kenya).}

\item{level}{A string specifying the level of analysis. Options are:
\itemize{
\item \code{'adminlevel_1'}: First administrative level.
\item \code{'district'}: District level.
}}
}
\value{
A \code{sf} object containing the shapefile for the specified country and level.
}
\description{
\code{get_country_shapefile} retrieves the shapefile for a specified country and
administrative level.
}
\examples{
shapefile <- get_country_shapefile('KEN', level = 'admin_level_1')

}
