% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{get_indicator_group_names}
\alias{get_indicator_group_names}
\title{Get Indicator Group Names}
\usage{
get_indicator_group_names()
}
\value{
Character vector of indicator group names
}
\description{
Get Indicator Group Names
}
