% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{get_indicator_groups}
\alias{get_indicator_groups}
\title{Get Indicator Groups}
\usage{
get_indicator_groups()
}
\value{
A named list of grouped indicators
}
\description{
Default grouping of indicators used in CD2030 coverage framework
}
