% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{get_named_indicators}
\alias{get_named_indicators}
\title{Get Named Indicator Vector}
\usage{
get_named_indicators()
}
\value{
Named character vector of indicators
}
\description{
Each indicator is named by its group
}
