% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{get_population_column}
\alias{get_population_column}
\title{Get Population Denominator Column Based on Indicator Only}
\usage{
get_population_column(indicator, denominator)
}
\arguments{
\item{indicator}{Character scalar. Indicator name (e.g., 'penta1', 'measles1')}

\item{denominator}{Character scalar. The denominator for the indicator}
}
\value{
A string naming the corresponding population column
}
\description{
Get Population Denominator Column Based on Indicator Only
}
