% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheConnection-class.R
\name{init_CacheConnection}
\alias{init_CacheConnection}
\title{Create a CacheConnection}
\usage{
init_CacheConnection(rds_path = NULL, countdown_data = NULL)
}
\arguments{
\item{rds_path}{Path to the RDS file (NULL for in-memory only).}

\item{countdown_data}{Countdown data (class \code{cd_data}).}
}
\value{
An object of class \link{CacheConnection}
}
\description{
Constructor function for objects of class \link{CacheConnection}.
}
