% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{list_tracer_vaccines}
\alias{list_tracer_vaccines}
\title{List Tracer Vaccines}
\usage{
list_tracer_vaccines()
}
\value{
A character vector of tracer vaccines
}
\description{
Returns a subset of vaccines used in tracer metrics
}
