% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{list_vaccines}
\alias{list_vaccines}
\title{List All Routine Vaccine Indicators}
\usage{
list_vaccines()
}
\value{
A character vector of vaccine names
}
\description{
Returns all vaccine indicators used in coverage estimates
}
