% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Processed Countdown 2030 Data from RDS or DTA File}
\usage{
load_data(path)
}
\arguments{
\item{path}{A string. The path to the file containing the processed data
in either RDS or DTA format.}
}
\value{
A tibble of class \code{cd_data} containing the loaded data.
}
\description{
\code{load_data} loads processed Countdown 2030 data from a specified file, either in
RDS or DTA format, and returns a tibble of class \code{cd_data}.
}
\details{
This function first checks the file extension to determine the format, then reads
the data accordingly. It supports both RDS and Stata (DTA) file formats commonly
used for saving processed datasets.
}
\examples{
\dontrun{
  # Load previously processed data from an RDS or DTA file
  data <- load_data('processed_cd2030_data.rds')
  data <- load_data('processed_cd2030_data.dta')
}

}
