% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey_data.R
\name{load_equity_data}
\alias{load_equity_data}
\title{Load and Process Country-Specific Survey Equity Data}
\usage{
load_equity_data(path)
}
\arguments{
\item{path}{Character. File path to the survey data file (DTA format).}
}
\value{
A tibble containing cleaned and processed survey data.
}
\description{
\code{load_equity_data} loads and processes survey data from a specified file path.
It performs data cleaning, renaming, and standardization while allowing for
the alignment of state codes.
}
