% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_excel_data}
\alias{load_excel_data}
\title{Load and Clean Countdown 2030 Excel Data}
\usage{
load_excel_data(
  filename,
  admin_data = "Admin_data",
  population_data = "Population_data",
  reporting_completeness = "Reporting_completeness",
  service_data = c("Service_data_1", "Service_data_2", "Service_data_3",
    "Vaccine_stock_data")
)
}
\arguments{
\item{filename}{A string. The path to the Excel file to be loaded.}

\item{admin_data}{A string. The name of the sheet containing administrative
data. Default is "Admin_data".}

\item{population_data}{A string. The name of the sheet containing population
data. Default is "Population_data".}

\item{reporting_completeness}{A string. The name of the sheet containing
reporting completeness data. Default is "Reporting_completeness".}

\item{service_data}{A vector of strings. The names of the sheets containing
service data. Default is \code{Service_data_1}, \code{Service_data_2}, \code{Service_data_3},
\code{Vaccine_stock_data}.}
}
\value{
A data of class \code{cd_data}.
}
\description{
This function loads and cleans data from an Excel file formatted for
Countdown 2030. It processes the provided Excel sheets, checks for
duplicates, merges data, and performs additional cleaning steps such as
fixing column names, removing unnecessary columns, and handling reporting rates.
}
\details{
The function performs several cleaning steps, including:
\itemize{
\item Removing header rows
\item Converting column types appropriately (e.g., converting characters to numeric)
\item Removing entirely missing rows and columns
\item Renaming columns (e.g., renaming "District" or "District_name" to "district")
\item Checking for duplicate entries based on key columns
\item Calculating reporting rates if necessary using expected and received values
\item Replacing special characters in the "Month" column and standardizing month names
}
}
\examples{
\dontrun{
  # Load and clean data from a Countdown 2030 Excel file
  data <- load_excel_data("countdown2030_data.xlsx")
}

}
