% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey_data.R
\name{load_un_estimates}
\alias{load_un_estimates}
\title{Load UN Estimates}
\usage{
load_un_estimates(path, country_iso, start_year, end_year)
}
\arguments{
\item{path}{Character. File path to the UN Estimates data file (DTA format).}

\item{country_iso}{Character. ISO code of the country.}

\item{start_year}{Integer. The year to start filtering the data}

\item{end_year}{Integer. The year to start filtering the data}
}
\value{
A tibble of class \code{cd_un_estimates} containing cleaned and processed
UN Estimate data.
}
\description{
This function loads and processes UN Estimate data from a specified file path.
It performs data cleaning, and renaming.
}
