% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.adjustment_values.R
\name{plot.cd_adjustment_values}
\alias{plot.cd_adjustment_values}
\title{Plot Adjusted vs. Unadjusted Data for Health Indicators}
\usage{
\method{plot}{cd_adjustment_values}(
  x,
  indicator = c("ideliv", "penta1", "bcg", "measles1", "anc1", "measles2", "opv1",
    "opv2", "opv3", "penta2", "penta3", "pcv1", "pcv2", "pcv3", "rota1", "rota2", "ipv1",
    "ipv2"),
  title = NULL,
  legend_labels = NULL,
  ...
)
}
\arguments{
\item{x}{A data frame containing the \code{year} column and columns for the raw
and adjusted values of health indicators (e.g., \code{ideliv_raw}, \code{ideliv_adj}).}

\item{indicator}{A character string specifying the prefix of the indicator to
plot (e.g., \code{"ideliv"}). Only the provided indicators will be plotted.}

\item{title}{A character string for the plot title. If \code{NULL}, a default title
based on the indicator is generated.}

\item{legend_labels}{A character vector of length 2 specifying custom labels
for the legend. The first element is used for the unadjusted (raw) data, and
the second element is for the adjusted data. If \code{NULL}, default labels
("N of \code{indicator} before adjustment" and "N of \code{indicator} after adjustment")
are generated.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A ggplot2 object showing the comparison of unadjusted and adjusted data
for the specified indicator over time.
}
\description{
\code{plot.cd_adjustment_values} creates a bar plot to compare the unadjusted (raw)
and adjusted values of health indicators over time. It allows users to specify
the indicator prefix and customize legend labels for flexibility across different
health data.
}
\details{
This function helps visualize the difference between raw and adjusted values
of a given health indicator, aiding in the assessment of data completeness and
adjustments. The difference and percentage difference are calculated within
the function but are not directly shown on the plot. Instead, the plot shows
the actual unadjusted and adjusted values side-by-side for each year.
}
\examples{
\dontrun{
  # Using default legend labels and title
  plot.cd_adjustment_values(adjustments, indicator = "ideliv")

  # Custom legend labels and title
  plot.cd_adjustment_values(adjustments, indicator = "instlivebirths",
                            title = "Customized Title",
                            legend_labels = c("Original", "Modified"))
}

}
