% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.completeness.R
\name{plot.cd_completeness_summary}
\alias{plot.cd_completeness_summary}
\title{Plot Missing Summary}
\usage{
\method{plot}{cd_completeness_summary}(x, indicator = NULL, ...)
}
\arguments{
\item{x}{A \code{cd_outlier} object containing pre-processed outlier data.}

\item{indicator}{Optional. One of the supported indicators (\code{"opv1"}, \code{"penta3"}, etc.)
to visualize in the plot. If \code{NULL} all indicators will be shown.}

\item{...}{Reserved for future use.}
}
\value{
A \code{ggplot} or \code{plotly} object depending on the selection type.
}
\description{
This method visualizes missing results for immunization indicators
}
\examples{
\dontrun{
  # Region-level summary
  plot(missing_data, indicator = "penta3")
}

}
