% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.indicator_coverage.R
\name{plot.cd_indicator_coverage}
\alias{plot.cd_indicator_coverage}
\title{Plot National Denominators and Coverage Indicators}
\usage{
\method{plot}{cd_indicator_coverage}(
  x,
  plot_type = c("anc_coverage_dhis2", "delivery_coverage_dhis2",
    "immunization_coverage_dhis2", "anc_coverage_un", "delivery_coverage_un",
    "immunization_coverage_un", "anc_coverage_anc1", "delivery_coverage_anc1",
    "immunization_coverage_anc1", "anc_coverage_penta1", "delivery_coverage_penta1",
    "immunization_coverage_penta1"),
  admin_name = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{cd_indicator_coverage} object containing the calculated indicators.}

\item{plot_type}{Character. Type of plot to generate. Options include:
\itemize{
\item \code{"population_dhis2"}: Population estimates from DHIS-2 data
\item \code{"births_dhis2"}: Live birth estimates from DHIS-2
\item \code{"population_un"}: Population estimates from UN data
\item \code{"births_un"}: Live birth estimates from UN
\item \code{"anc_coverage"}: ANC coverage indicators
\item \code{"immunization_coverage_dhis2"}: DHIS-2 based immunization coverage indicators
\item \code{"immunization_coverage_un"}: UN-based immunization coverage
\item \code{"immunization_coverage_anc1"}: ANC-1 based immunization coverage
}}

\item{admin_name}{Name of the area to plot}

\item{...}{Additional arguments (currently not used).}
}
\value{
A ggplot object for the selected plot type.
}
\description{
Generates specific plots for national denominators and health coverage indicators based on the
calculated denominators in the \code{cd_indicator_coverage} object.
}
