% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mapping.R
\name{plot.cd_mapping}
\alias{plot.cd_mapping}
\title{Plot Subnational Mapping Data}
\usage{
\method{plot}{cd_mapping}(
  x,
  indicator = NULL,
  denominator = c("dhis2", "anc1", "penta1", "penta1derived"),
  palette = c("Reds", "Blues", "Greens", "Purples", "YlGnBu"),
  plot_year = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{A data frame containing the subnational mapping data.}

\item{indicator}{A string specifying the indicator to visualize. Options include:
\itemize{
\item \code{"anc1"}
\item \code{"bcg"}
\item \code{"measles1"}
\item \code{"measles2"}
\item \code{"measles3"}
\item \code{"opv1"}
\item \code{"opv2"}
\item \code{"opv3"}
\item \code{"pcv1"}
\item \code{"pcv2"}
\item \code{"pcv3"}
\item \code{"penta1"}
\item \code{"penta2"}
\item \code{"penta2"}
\item \code{"rota1"}
\item \code{"rota2"}
\item \code{"instdeliveries"}
\item \code{"ipv1"}
\item \code{"ipv2"}
\item \code{"undervax"}
\item \code{"zerodose"}
\item \code{"dropout_measles12"}
\item \code{"dropout_penta13"}
\item \code{"dropout_penta3mcv1"}
}}

\item{denominator}{A string specifying the denominator used in the calculation. Options are:
\itemize{
\item \code{"dhis2"}
\item \code{"anc1"}
\item \code{"penta1"}
}}

\item{palette}{A string specifying the color palette. Options include:
\itemize{
\item \code{"Reds"}
\item \code{"Blues"}
\item \code{"Greens"}
\item \code{"Purples"}
\item \code{"YlGnBu"}
}}

\item{plot_year}{Optional. A numeric value or vector specifying the year(s) to
include in the plot. If \code{NULL}, all years in the data are included.}

\item{title}{Optional. A string specifying the plot title. Defaults to
\code{"Distribution of [indicator] by Regions"}.}

\item{...}{Other params to pass to plot}
}
\value{
A \code{ggplot} object representing the subnational mapping plot.
}
\description{
This function generates a subnational mapping plot for a specified indicator,
using a gradient color scheme.
}
\examples{
\dontrun{
  # Plot Penta1 coverage with DHIS2 denominator
  plot.cd_mapping(
    x = data,
    indicator = "penta1",
    denominator = "dhis2",
    palette = "Blues",
    plot_year = 2022
  )
}

}
