% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.national_coverage.R
\name{plot.cd_national_coverage}
\alias{plot.cd_national_coverage}
\title{Plot National Coverage Data}
\usage{
\method{plot}{cd_national_coverage}(x, ...)
}
\arguments{
\item{x}{A data frame of type \code{cd_national_coverage}, containing year-wise
national coverage data for the specified country and indicator.}

\item{...}{Additional arguments passed to or from other methods (currently
unused).}
}
\value{
A ggplot object displaying a line plot of the national coverage data
by year, with lines and points representing different estimate types.
}
\description{
This function generates a line plot to visualize national immunization coverage
data across years, allowing comparison between different estimates (e.g., DHIS2
estimates, WUENIC estimates, and Survey estimates).
}
\details{
This plot function transforms the coverage data into a long format suitable for
plotting with ggplot2, and extracts the denominator information for display in
the plot caption. It shows coverage values for each year, with different lines
and points for each estimate type.

The plot includes:
\itemize{
\item \strong{Y-axis}: Coverage percentage.
\item \strong{X-axis}: Year.
\item \strong{Line color}: Represents different estimates (e.g., DHIS2, Survey).
\item \strong{Caption}: Indicates the denominator used in the coverage calculation.
}

This function is meant to be used with the \code{cd_national_coverage} class, which
contains coverage values for different indicators and estimates.
}
\examples{
\dontrun{
  plot(analyze_national_coverage(dt_adj, indicator = "bcg", denominator = "anc1"))
}

}
