% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.population_metrics.R
\name{plot.cd_population_metrics}
\alias{plot.cd_population_metrics}
\title{Plot National Population or Births Metrics}
\usage{
\method{plot}{cd_population_metrics}(x, metric = c("population", "births"), ...)
}
\arguments{
\item{x}{A \code{cd_population_metrics} object containing national-level demographic
data.}

\item{metric}{A character string specifying the type of data to plot. It must
be one of:
\itemize{
\item \strong{'population'}: Plots total population estimates in thousands from DHIS-2 and UN sources.
\item \strong{'births'}: Plots total live births in thousands from DHIS-2 and UN sources.
}}

\item{...}{Additional arguments (not used in this function).}
}
\value{
A ggplot object visualizing the selected demographic data over time.
}
\description{
This function generates a plot to visualize national-level demographic data
for total population or live births over time.It compares DHIS-2 and UN
projections for the specified metric, limited to national-level data due to
the availability of UN estimates.
}
\details{
The \code{plot_population_metrics} function provides visualizations based on the
selected \code{metric}:
\itemize{
\item \strong{Population Plot}: Shows total population (in thousands) by year, with
separate lines for DHIS-2 and UN projections.
\item \strong{Live Births Plot}: Displays total live births (in thousands) by year,
comparing DHIS-2 and UN projections.
}

This function supports only national-level data as UN estimates are not available
at subnational levels. Attempting to plot subnational data will result in an error.
}
\examples{
\dontrun{
  # Plot total population estimates at the national level
  plot(population_metrics_data, metric = 'population')

  # Plot total live births estimates at the national level
  plot(population_metrics_data, metric = 'births')
}

}
