% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absolute_differences.R
\name{plot_absolute_differences}
\alias{plot_absolute_differences}
\title{Plot DHIS2-Based Indicator Coverage with Different Denominators and Survey Coverage}
\usage{
plot_absolute_differences(
  .data,
  indicator = c("penta3", "measles1", "bcg"),
  survey_coverage = 88
)
}
\arguments{
\item{.data}{A data frame containing columns for \code{country}, \code{year}, and the coverage values
for various denominators. Each indicator should have columns ending with suffixes like \verb{_dhis2}, \verb{_anc1}, \verb{_penta1}, and \verb{_un}.}

\item{indicator}{A character string specifying the health indicator to plot. Options include, but are not limited to:
\code{mcv1}, \code{penta3}, \code{bcg}.}

\item{survey_coverage}{A numeric value representing the national survey coverage percentage
to be displayed as a reference line on the plot. Default is \code{88}.}
}
\value{
A \code{ggplot2} object showing the coverage percentages for the selected indicator
from various denominator sources, with a reference line for survey coverage.
}
\description{
\code{plot_absolute_differences} generates a bar plot comparing facility-based coverage
percentages for a selected health indicator with different denominator sources (e.g., DHIS2,
ANC1-derived, Penta1-derived, UN projection) and overlays a horizontal line representing
national survey coverage for the specified year.
}
\details{
This function allows for visualization of the discrepancy between facility-based coverage
data sourced from different denominators (e.g., DHIS2 projections, ANC1-derived, Penta1-derived,
UN projections) and the national survey coverage. It helps in comparing how coverage estimates vary
depending on the denominator, aiding in the assessment of data completeness and consistency.
}
\examples{
\dontrun{
  # Plot for the "penta3" indicator in the year 2021 with a survey coverage of 90
  plot.cd_indicator_coverage(data, indicator = "penta3", survey_coverage = 90, year = 2021)

  # Plot for the "bcg" indicator with default survey coverage and year 2020
  plot.cd_indicator_coverage(data, indicator = "bcg", year = 2020)
}

}
