% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_plot_comparison.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Generic Function for Plotting Comparisons}
\usage{
plot_comparison(
  .data,
  x_var,
  y_var,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  show_r_squared = TRUE,
  facet_ncol = 3,
  plot_theme = theme_bw(),
  call = caller_env(),
  ...
)
}
\arguments{
\item{.data}{A data object to be plotted.}

\item{x_var}{A string specifying the column name for the x-axis variable (required for \code{cd_data} objects).}

\item{y_var}{A string specifying the column name for the y-axis variable (required for \code{cd_data} objects).}

\item{title}{An optional title for the plot.}

\item{x_label}{An optional label for the x-axis.}

\item{y_label}{An optional label for the y-axis.}

\item{show_r_squared}{Logical; whether to display R-squared values on each facet. Default is TRUE.}

\item{facet_ncol}{Number of columns for faceting by year. Default is 3.}

\item{plot_theme}{ggplot theme to apply to the plot. Default is \code{theme_bw()}.}

\item{call}{The caller environment}

\item{...}{Additional arguments passed to specific methods, such as \code{plot_comparison.cd_data}.}
}
\description{
\code{plot_comparison} is a generic function that creates a comparison plot for
different data objects. It will dispatch to specific methods depending on
the class of the input object.
}
