% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{plot_comparison_ipv1_ipv2}
\alias{plot_comparison_ipv1_ipv2}
\title{Plot Comparison of IPV1 and IPV2}
\usage{
plot_comparison_ipv1_ipv2(data)
}
\arguments{
\item{data}{A data frame containing IPV1 and IPV2 data.}
}
\value{
A ggplot object comparing IPV1 and IPV2
}
\description{
This function compares IPV1 and IPV2 indicators across different years
using a scatter plot with a linear fit.
}
\examples{
\dontrun{

  plot_comparison_ipv1_ipv2(data)
}
}
