% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{plot_comparison_opv1_anc1}
\alias{plot_comparison_opv1_anc1}
\title{Plot Comparison of OPV1 and ANC1}
\usage{
plot_comparison_opv1_anc1(data)
}
\arguments{
\item{data}{A data frame containing OPV1 and ANC1 data.}
}
\value{
A ggplot object comparing OPV1 and ANC1
}
\description{
This function compares OPV1 and ANC1 indicators across different years
using a scatter plot with a linear fit.
}
\examples{
\dontrun{

  plot_comparison_opv1_anc1(data)
}
}
