% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_plot_comparison.R
\name{plot_comparison_opv1_opv3}
\alias{plot_comparison_opv1_opv3}
\title{Plot Comparison of OPV1 and OPV3}
\usage{
plot_comparison_opv1_opv3(.data)
}
\arguments{
\item{.data}{A data frame containing OPV1 and OPV3 data.}
}
\value{
A ggplot object comparing OPV1 and OPV3
}
\description{
This function compares OPV1 and OPV3 indicators across different years
using a scatter plot with a linear fit.
}
\examples{
\dontrun{

  plot_comparison_opv1_opv3(data)
}
}
