% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_plot_comparison.R
\name{plot_comparison_opv3_penta3}
\alias{plot_comparison_opv3_penta3}
\title{Plot Comparison of OPV3 and Penta3}
\usage{
plot_comparison_opv3_penta3(.data)
}
\arguments{
\item{.data}{A data frame containing OPV3 and Penta3 data.}
}
\value{
A ggplot object comparing OPV3 and Penta3
}
\description{
This function compares OPV3 and Penta3 indicators across different years
using a scatter plot with a linear fit.
}
\examples{
\dontrun{

  plot_comparison_opv3_penta3(data)
}
}
