% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_plot_comparison.R
\name{plot_comparison_pcv1_pcv3}
\alias{plot_comparison_pcv1_pcv3}
\title{Plot Comparison of PCV1 and PCV3}
\usage{
plot_comparison_pcv1_pcv3(.data)
}
\arguments{
\item{.data}{A data frame containing PCV1 and PCV3 data.}
}
\value{
A ggplot object comparing PCV1 and PCV3
}
\description{
This function compares PCV1 and PCV3 indicators across different years
using a scatter plot with a linear fit.
}
\examples{
\dontrun{

  plot_comparison_pcv1_pcv3(data)
}
}
