% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_plot_comparison.R
\name{plot_comparison_rota1_rota2}
\alias{plot_comparison_rota1_rota2}
\title{Plot Comparison of ROTA1 and ROTA2}
\usage{
plot_comparison_rota1_rota2(.data)
}
\arguments{
\item{.data}{A data frame containing ROTA1 and ROTA2 data.}
}
\value{
A ggplot object comparing ROTA1 and ROTA2
}
\description{
This function compares ROTA1 and ROTA2 indicators across different years
using a scatter plot with a linear fit.
}
\examples{
\dontrun{

  plot_comparison_rota1_rota2(data)
}
}
