% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_line_graph}
\alias{plot_line_graph}
\title{Plot Line Graph for Multiple Series with Dynamic Y-axis Scaling}
\usage{
plot_line_graph(
  .data,
  x,
  y_vars,
  y_labels,
  title,
  y_axis_title,
  hline = NULL,
  hline_style = "dashed"
)
}
\arguments{
\item{.data}{A data frame containing the variables to plot.}

\item{x}{The unquoted column name for the x-axis variable (e.g., \code{year}).}

\item{y_vars}{A character vector of column names for the y variables to plot.}

\item{y_labels}{A character vector of labels for the y variables (must match the length of \code{y_vars}).}

\item{title}{The title of the plot.}

\item{y_axis_title}{The title of the y-axis.}

\item{hline}{An optional numeric value to draw a horizontal line.}

\item{hline_style}{The line style for the horizontal line, default is "dashed".}
}
\value{
A ggplot object.
}
\description{
Generates a line graph for specified y variables over a shared x-axis, with dynamic scaling of the y-axis based on the data.
}
