% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_reporting.R
\name{print_reporting_rate}
\alias{print_reporting_rate}
\title{Plot Reporting Rate Summary}
\usage{
print_reporting_rate(data)
}
\arguments{
\item{data}{A data frame containing reporting rate data that has already been
processed by \code{\link[=check_reporting_rates]{check_reporting_rates()}}.}
}
\value{
A ggplot object displaying reporting rates across indicators and years.
}
\description{
This function creates a bar plot of reporting rates, showing the percentage
of districts with reporting rates for different indicators over the years.
}
\examples{
\dontrun{

  print_reporting_rate(data)
}
}
