% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_years.R
\name{remove_years}
\alias{remove_years}
\title{Remove Specified Years from DHIS-2 Dataset}
\usage{
remove_years(.data, years_to_drop)
}
\arguments{
\item{.data}{A dataframe containing DHIS-2 data with a \code{year} column, representing
demographic information by year.}

\item{years_to_drop}{An integer vector specifying the years to be removed from the dataset.}
}
\value{
A data of class \code{cd_data}, but without rows for the specified years in \code{years_to_drop}.
}
\description{
This function filters out specified years from a DHIS-2 dataset.
}
\details{
The function:
\enumerate{
\item Verifies that \code{.data} is a \code{cd_data} count down data.
\item Checks that \code{years_to_drop} is an integer vector.
\itemize{
\item If \code{years_to_drop} is not a vector of integers, an error is raised.
}
\item Filters \code{.data} to exclude rows where \code{year} matches any value in \code{years_to_drop}.
}
}
\examples{
\dontrun{
  # Example usage to remove data for 2017 and 2018:
  filtered_data <- remove_years(dhis_data, years_to_drop = c(2017, 2018))
}

}
