% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_reporting_rate.R
\name{tbl_sum.cd_average_reporting_rate}
\alias{tbl_sum.cd_average_reporting_rate}
\title{Summary for cd_average_reporting_rate}
\usage{
\method{tbl_sum}{cd_average_reporting_rate}(x, ...)
}
\arguments{
\item{x}{A \code{cd_average_reporting_rate} object containing reporting rate data.}

\item{...}{Additional arguments for compatibility with S3 method.}
}
\value{
A character vector summarizing the content and purpose of the data.
}
\description{
Provides a custom summary for the \code{cd_average_reporting_rate} object, displaying
a message indicating the data shows average reporting rates by year for all indicators.
}
