% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_ratio.R
\name{tbl_sum.cd_check_ratios}
\alias{tbl_sum.cd_check_ratios}
\title{Summary for cd_check_ratios}
\usage{
\method{tbl_sum}{cd_check_ratios}(x, ...)
}
\arguments{
\item{x}{A \code{cd_check_ratios} object containing reporting rate data.}

\item{...}{Additional arguments for compatibility with S3 method.}
}
\value{
A character vector summarizing the content and purpose of the data.
}
\description{
Provides a custom summary for the \code{cd_check_ratios} object, displaying
a message indicating the data shows Percentage of districts with adequate ratios
by year for all indicators.
}
