% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_outlier.R
\name{tbl_sum.cd_extreme_outliers}
\alias{tbl_sum.cd_extreme_outliers}
\title{Summary for \code{cd_extreme_outliers}}
\usage{
\method{tbl_sum}{cd_extreme_outliers}(x, ...)
}
\arguments{
\item{x}{A \code{cd_extreme_outliers} object containing the percentage of districts without extreme outliers
by year and indicator.}

\item{...}{Additional arguments for compatibility with S3 methods.}
}
\value{
A character vector describing the purpose of the data, highlighting the percentage of districts
with no extreme outliers for each year.
}
\description{
Provides a summary of the \code{cd_extreme_outliers} object, indicating the percentage of districts with no
extreme outliers by year for each indicator. This summary facilitates data quality monitoring by showing
the proportion of monthly values that are non-extreme for all indicators.
}
