% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_outlier.R
\name{tbl_sum.cd_outliers_summary}
\alias{tbl_sum.cd_outliers_summary}
\title{Summary for \code{cd_outliers_summary}}
\usage{
\method{tbl_sum}{cd_outliers_summary}(x, ...)
}
\arguments{
\item{x}{A \code{cd_outliers_summary} object containing data quality metrics.}

\item{...}{Additional arguments for compatibility with S3 methods.}
}
\value{
A character vector summarizing the content and purpose of the data,
indicating the proportion of values that meet the criteria for non-outliers.
}
\description{
Provides a custom summary for the \code{cd_outliers_summary} object, displaying
the percentage of monthly indicator values that are not considered extreme outliers, aggregated by year.
}
