% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_inequality.R
\name{analyze_inequality}
\alias{analyze_inequality}
\title{Analyze Subnational Health Coverage Data with Disparity Metrics}
\usage{
analyze_inequality(
  .data,
  level = c("admin_level_1", "district"),
  indicator = c("bcg", "anc1", "pcv3", "opv1", "opv2", "opv3", "penta2", "pcv1", "pcv2",
    "penta1", "penta3", "measles1", "rota1", "rota2", "instdeliveries", "measles2",
    "ipv1", "ipv2", "undervax", "dropout_penta13", "zerodose", "dropout_measles12",
    "dropout_penta3mcv1"),
  denominator = c("dhis2", "anc1", "penta1"),
  un_estimates,
  sbr = 0.02,
  nmr = 0.025,
  pnmr = 0.024,
  anc1survey = 0.98,
  dpt1survey = 0.97,
  twin = 0.015,
  preg_loss = 0.03
)
}
\arguments{
\item{.data}{A data frame containing the health coverage data.}

\item{level}{A character string specifying the administrative level to analyze
(either "admin_level_1" or "district").}

\item{indicator}{A character string specifying the health indicator to analyze
(e.g., "bcg", "measles1").}

\item{denominator}{A character string specifying the denominator to use for
coverage calculations (e.g., "dhis2", "anc1", "penta1").}
}
\value{
A \code{cd_inequality} object containing calculated metrics such
as coverage, MADM, MRDM, and population shares.
}
\description{
\code{analyze_inequality} Calculates health coverage and disparity metrics,
such as Mean Absolute Difference to the Mean (MADM), weighted MADM, Mean Relative
Difference to the Mean (MRDM), and weighted MRDM, for a specified subnational
level and health indicator. The function returns a \code{cd_inequality}
object with calculated metrics for plotting and summarizing.
}
\examples{
\dontrun{
data <- analyze_inequality(.data, level = "district",
                           indicator = "measles1", denominator = "penta1")
}

}
