% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheConnection-class.R
\name{CacheConnection}
\alias{CacheConnection}
\title{CacheConnection R6 Class}
\description{
An R6 class for handling RDS caching for data analysis and report generation.
}
\keyword{CacheConnection}
\keyword{Initialize}
\keyword{class}
\keyword{internal}
\keyword{the}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CacheConnection-new}{\code{CacheConnection$new()}}
\item \href{#method-CacheConnection-load_from_disk}{\code{CacheConnection$load_from_disk()}}
\item \href{#method-CacheConnection-save_to_disk}{\code{CacheConnection$save_to_disk()}}
\item \href{#method-CacheConnection-append_page_note}{\code{CacheConnection$append_page_note()}}
\item \href{#method-CacheConnection-get_notes}{\code{CacheConnection$get_notes()}}
\item \href{#method-CacheConnection-reactive}{\code{CacheConnection$reactive()}}
\item \href{#method-CacheConnection-set_cache_path}{\code{CacheConnection$set_cache_path()}}
\item \href{#method-CacheConnection-set_countdown_data}{\code{CacheConnection$set_countdown_data()}}
\item \href{#method-CacheConnection-set_adjusted_data}{\code{CacheConnection$set_adjusted_data()}}
\item \href{#method-CacheConnection-set_performance_threshold}{\code{CacheConnection$set_performance_threshold()}}
\item \href{#method-CacheConnection-set_excluded_years}{\code{CacheConnection$set_excluded_years()}}
\item \href{#method-CacheConnection-set_k_factors}{\code{CacheConnection$set_k_factors()}}
\item \href{#method-CacheConnection-set_adjusted_flag}{\code{CacheConnection$set_adjusted_flag()}}
\item \href{#method-CacheConnection-set_survey_estimates}{\code{CacheConnection$set_survey_estimates()}}
\item \href{#method-CacheConnection-set_national_estimates}{\code{CacheConnection$set_national_estimates()}}
\item \href{#method-CacheConnection-set_start_survey_year}{\code{CacheConnection$set_start_survey_year()}}
\item \href{#method-CacheConnection-set_denominator}{\code{CacheConnection$set_denominator()}}
\item \href{#method-CacheConnection-set_selected_admin_level_1}{\code{CacheConnection$set_selected_admin_level_1()}}
\item \href{#method-CacheConnection-set_selected_district}{\code{CacheConnection$set_selected_district()}}
\item \href{#method-CacheConnection-set_mapping_years}{\code{CacheConnection$set_mapping_years()}}
\item \href{#method-CacheConnection-set_un_estimates}{\code{CacheConnection$set_un_estimates()}}
\item \href{#method-CacheConnection-set_wuenic_estimates}{\code{CacheConnection$set_wuenic_estimates()}}
\item \href{#method-CacheConnection-set_national_survey}{\code{CacheConnection$set_national_survey()}}
\item \href{#method-CacheConnection-set_regional_survey}{\code{CacheConnection$set_regional_survey()}}
\item \href{#method-CacheConnection-set_wiq_survey}{\code{CacheConnection$set_wiq_survey()}}
\item \href{#method-CacheConnection-set_area_survey}{\code{CacheConnection$set_area_survey()}}
\item \href{#method-CacheConnection-set_education_survey}{\code{CacheConnection$set_education_survey()}}
\item \href{#method-CacheConnection-set_survey_mapping}{\code{CacheConnection$set_survey_mapping()}}
\item \href{#method-CacheConnection-set_map_mapping}{\code{CacheConnection$set_map_mapping()}}
\item \href{#method-CacheConnection-clone}{\code{CacheConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$new(rds_path = NULL, countdown_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rds_path}}{Path to the RDS file (NULL for in-memory only).}

\item{\code{countdown_data}}{Countdown data (class \code{cd_data}).
Load data from RDS file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-load_from_disk"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-load_from_disk}{}}}
\subsection{Method \code{load_from_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$load_from_disk()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None
Save data to disk (only if changed and RDS path is not NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-save_to_disk"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-save_to_disk}{}}}
\subsection{Method \code{save_to_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$save_to_disk()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None
Reactive method for the CacheConnection object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-append_page_note"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-append_page_note}{}}}
\subsection{Method \code{append_page_note()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$append_page_note(
  page_id,
  object_id,
  note,
  parameters = list(),
  include_in_report = FALSE,
  include_plot_table = FALSE,
  single_entry = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-get_notes"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-get_notes}{}}}
\subsection{Method \code{get_notes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$get_notes(page_id, object_id = NULL, parameters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-reactive"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-reactive}{}}}
\subsection{Method \code{reactive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$reactive()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A Shiny reactive expression tracking changes
Update a field (with change tracking)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_cache_path"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_cache_path}{}}}
\subsection{Method \code{set_cache_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_cache_path(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_countdown_data"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_countdown_data}{}}}
\subsection{Method \code{set_countdown_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_countdown_data(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_adjusted_data"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_adjusted_data}{}}}
\subsection{Method \code{set_adjusted_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_adjusted_data(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_performance_threshold"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_performance_threshold}{}}}
\subsection{Method \code{set_performance_threshold()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_performance_threshold(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_excluded_years"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_excluded_years}{}}}
\subsection{Method \code{set_excluded_years()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_excluded_years(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_k_factors"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_k_factors}{}}}
\subsection{Method \code{set_k_factors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_k_factors(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_adjusted_flag"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_adjusted_flag}{}}}
\subsection{Method \code{set_adjusted_flag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_adjusted_flag(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_survey_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_survey_estimates}{}}}
\subsection{Method \code{set_survey_estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_survey_estimates(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_national_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_national_estimates}{}}}
\subsection{Method \code{set_national_estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_national_estimates(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_start_survey_year"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_start_survey_year}{}}}
\subsection{Method \code{set_start_survey_year()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_start_survey_year(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_denominator"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_denominator}{}}}
\subsection{Method \code{set_denominator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_denominator(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_selected_admin_level_1"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_selected_admin_level_1}{}}}
\subsection{Method \code{set_selected_admin_level_1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_selected_admin_level_1(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_selected_district"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_selected_district}{}}}
\subsection{Method \code{set_selected_district()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_selected_district(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_mapping_years"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_mapping_years}{}}}
\subsection{Method \code{set_mapping_years()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_mapping_years(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_un_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_un_estimates}{}}}
\subsection{Method \code{set_un_estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_un_estimates(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_wuenic_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_wuenic_estimates}{}}}
\subsection{Method \code{set_wuenic_estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_wuenic_estimates(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_national_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_national_survey}{}}}
\subsection{Method \code{set_national_survey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_national_survey(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_regional_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_regional_survey}{}}}
\subsection{Method \code{set_regional_survey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_regional_survey(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_wiq_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_wiq_survey}{}}}
\subsection{Method \code{set_wiq_survey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_wiq_survey(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_area_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_area_survey}{}}}
\subsection{Method \code{set_area_survey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_area_survey(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_education_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_education_survey}{}}}
\subsection{Method \code{set_education_survey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_education_survey(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_survey_mapping"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_survey_mapping}{}}}
\subsection{Method \code{set_survey_mapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_survey_mapping(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_map_mapping"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_map_mapping}{}}}
\subsection{Method \code{set_map_mapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_map_mapping(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
