% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheConnection-class.R
\docType{class}
\name{CacheConnection}
\alias{CacheConnection}
\title{CacheConnection Class}
\format{
An \link[R6:R6Class]{R6::R6Class} generator object.
}
\description{
An R6 class that handles persistent or in-memory caching of data used in the Countdown 2030
analysis and reporting workflows. It supports tracking of various internal data objects,
reactive updates (for Shiny apps), note-taking for report annotations, and saving/loading
from \code{.rds} files.
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{language}}{Get the UI language.}

\item{\code{cache_path}}{Get cache path.}

\item{\code{countdown_data}}{Get countdown data.}

\item{\code{country}}{Get country name.}

\item{\code{country_iso}}{Get country ISO3 code.}

\item{\code{default_national_estimates}}{Get the national rates.}

\item{\code{adjusted_data}}{Gets adjusted data.}

\item{\code{data_with_excluded_years}}{Get data with excluded years removed.}

\item{\code{performance_threshold}}{Gets performance threshold.}

\item{\code{excluded_years}}{Gets excluded years.}

\item{\code{k_factors}}{Gets k-factors.}

\item{\code{adjusted_flag}}{Gets adjusted flag.}

\item{\code{survey_estimates}}{Gets survey estimates.}

\item{\code{national_estimates}}{Gets national estimates.}

\item{\code{survey_years}}{Get survey years.}

\item{\code{survey_source}}{Gets survey source of information.}

\item{\code{survey_year}}{Gets survey year of survey estimates.}

\item{\code{start_survey_year}}{Gets start survey year.}

\item{\code{denominator}}{Gets denominator.}

\item{\code{selected_admin_level_1}}{Gets selected region.}

\item{\code{selected_district}}{Gets selected district.}

\item{\code{mapping_years}}{Gets mapping years.}

\item{\code{un_estimates}}{Gets UN estimates.}

\item{\code{wuenic_estimates}}{Gets WUENIC estimates.}

\item{\code{national_survey}}{Gets national survey.}

\item{\code{regional_survey}}{Gets regional survey.}

\item{\code{wiq_survey}}{Gets WIQ survey.}

\item{\code{area_survey}}{Gets area survey.}

\item{\code{education_survey}}{Gets  education survey.}

\item{\code{survey_mapping}}{Gets survey mapping.}

\item{\code{map_mapping}}{Gets map mapping.
Update a field (with change tracking)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CacheConnection-new}{\code{CacheConnection$new()}}
\item \href{#method-CacheConnection-load_from_disk}{\code{CacheConnection$load_from_disk()}}
\item \href{#method-CacheConnection-save_to_disk}{\code{CacheConnection$save_to_disk()}}
\item \href{#method-CacheConnection-append_page_note}{\code{CacheConnection$append_page_note()}}
\item \href{#method-CacheConnection-get_notes}{\code{CacheConnection$get_notes()}}
\item \href{#method-CacheConnection-reactive}{\code{CacheConnection$reactive()}}
\item \href{#method-CacheConnection-set_language}{\code{CacheConnection$set_language()}}
\item \href{#method-CacheConnection-set_cache_path}{\code{CacheConnection$set_cache_path()}}
\item \href{#method-CacheConnection-set_countdown_data}{\code{CacheConnection$set_countdown_data()}}
\item \href{#method-CacheConnection-set_adjusted_data}{\code{CacheConnection$set_adjusted_data()}}
\item \href{#method-CacheConnection-set_performance_threshold}{\code{CacheConnection$set_performance_threshold()}}
\item \href{#method-CacheConnection-set_excluded_years}{\code{CacheConnection$set_excluded_years()}}
\item \href{#method-CacheConnection-set_k_factors}{\code{CacheConnection$set_k_factors()}}
\item \href{#method-CacheConnection-set_adjusted_flag}{\code{CacheConnection$set_adjusted_flag()}}
\item \href{#method-CacheConnection-set_survey_estimates}{\code{CacheConnection$set_survey_estimates()}}
\item \href{#method-CacheConnection-set_national_estimates}{\code{CacheConnection$set_national_estimates()}}
\item \href{#method-CacheConnection-set_survey_source}{\code{CacheConnection$set_survey_source()}}
\item \href{#method-CacheConnection-set_survey_year}{\code{CacheConnection$set_survey_year()}}
\item \href{#method-CacheConnection-set_start_survey_year}{\code{CacheConnection$set_start_survey_year()}}
\item \href{#method-CacheConnection-set_denominator}{\code{CacheConnection$set_denominator()}}
\item \href{#method-CacheConnection-set_selected_admin_level_1}{\code{CacheConnection$set_selected_admin_level_1()}}
\item \href{#method-CacheConnection-set_selected_district}{\code{CacheConnection$set_selected_district()}}
\item \href{#method-CacheConnection-set_mapping_years}{\code{CacheConnection$set_mapping_years()}}
\item \href{#method-CacheConnection-set_un_estimates}{\code{CacheConnection$set_un_estimates()}}
\item \href{#method-CacheConnection-set_wuenic_estimates}{\code{CacheConnection$set_wuenic_estimates()}}
\item \href{#method-CacheConnection-set_national_survey}{\code{CacheConnection$set_national_survey()}}
\item \href{#method-CacheConnection-set_regional_survey}{\code{CacheConnection$set_regional_survey()}}
\item \href{#method-CacheConnection-set_wiq_survey}{\code{CacheConnection$set_wiq_survey()}}
\item \href{#method-CacheConnection-set_area_survey}{\code{CacheConnection$set_area_survey()}}
\item \href{#method-CacheConnection-set_education_survey}{\code{CacheConnection$set_education_survey()}}
\item \href{#method-CacheConnection-set_survey_mapping}{\code{CacheConnection$set_survey_mapping()}}
\item \href{#method-CacheConnection-set_map_mapping}{\code{CacheConnection$set_map_mapping()}}
\item \href{#method-CacheConnection-clone}{\code{CacheConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-new}{}}}
\subsection{Method \code{new()}}{
Initialize a CacheConnection instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$new(rds_path = NULL, countdown_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rds_path}}{Path to the RDS file (can be NULL).}

\item{\code{countdown_data}}{Countdown data of class \code{cd_data}.
Load data from disk.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-load_from_disk"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-load_from_disk}{}}}
\subsection{Method \code{load_from_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$load_from_disk()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None. Updates internal state.
Save data to disk (only if changed and RDS path is not NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-save_to_disk"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-save_to_disk}{}}}
\subsection{Method \code{save_to_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$save_to_disk()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None. Updates file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-append_page_note"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-append_page_note}{}}}
\subsection{Method \code{append_page_note()}}{
Append a note to a page.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$append_page_note(
  page_id,
  object_id,
  note,
  parameters = list(),
  include_in_report = FALSE,
  include_plot_table = FALSE,
  single_entry = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{page_id}}{Page ID.}

\item{\code{object_id}}{Object ID.}

\item{\code{note}}{Note text.}

\item{\code{parameters}}{Named list of parameters.}

\item{\code{include_in_report}}{Logical flag.}

\item{\code{include_plot_table}}{Logical flag.}

\item{\code{single_entry}}{Logical flag for uniqueness.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-get_notes"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-get_notes}{}}}
\subsection{Method \code{get_notes()}}{
Retrieve notes for a given page/object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$get_notes(page_id, object_id = NULL, parameters = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{page_id}}{Page ID.}

\item{\code{object_id}}{Object ID.}

\item{\code{parameters}}{Named list of parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-reactive"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-reactive}{}}}
\subsection{Method \code{reactive()}}{
Return a reactive wrapper (for Shiny).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$reactive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_language"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_language}{}}}
\subsection{Method \code{set_language()}}{
Set the language.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_language(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A scalar string representing 2 digit string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_cache_path"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_cache_path}{}}}
\subsection{Method \code{set_cache_path()}}{
Set cache RDS file path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_cache_path(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{New file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_countdown_data"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_countdown_data}{}}}
\subsection{Method \code{set_countdown_data()}}{
Set the countdown data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_countdown_data(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A \code{cd_data} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_adjusted_data"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_adjusted_data}{}}}
\subsection{Method \code{set_adjusted_data()}}{
Set adjusted data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_adjusted_data(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A \code{cd_data} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_performance_threshold"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_performance_threshold}{}}}
\subsection{Method \code{set_performance_threshold()}}{
Set performance threshold.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_performance_threshold(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numeric scalar.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_excluded_years"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_excluded_years}{}}}
\subsection{Method \code{set_excluded_years()}}{
Set years to exclude.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_excluded_years(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Numeric vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_k_factors"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_k_factors}{}}}
\subsection{Method \code{set_k_factors()}}{
Set K-factors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_k_factors(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Named numeric vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_adjusted_flag"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_adjusted_flag}{}}}
\subsection{Method \code{set_adjusted_flag()}}{
Set adjusted flag.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_adjusted_flag(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Logical scalar.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_survey_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_survey_estimates}{}}}
\subsection{Method \code{set_survey_estimates()}}{
Set survey estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_survey_estimates(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Named numeric vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_national_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_national_estimates}{}}}
\subsection{Method \code{set_national_estimates()}}{
Set national estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_national_estimates(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Named list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_survey_source"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_survey_source}{}}}
\subsection{Method \code{set_survey_source()}}{
Set year of survey estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_survey_source(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character scalar.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_survey_year"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_survey_year}{}}}
\subsection{Method \code{set_survey_year()}}{
Set year of survey estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_survey_year(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Integer year.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_start_survey_year"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_start_survey_year}{}}}
\subsection{Method \code{set_start_survey_year()}}{
Set start year of surveys.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_start_survey_year(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Integer year.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_denominator"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_denominator}{}}}
\subsection{Method \code{set_denominator()}}{
Set denominator type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_denominator(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character scalar.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_selected_admin_level_1"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_selected_admin_level_1}{}}}
\subsection{Method \code{set_selected_admin_level_1()}}{
Set selected region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_selected_admin_level_1(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character scalar.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_selected_district"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_selected_district}{}}}
\subsection{Method \code{set_selected_district()}}{
Set selected district.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_selected_district(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character scalar.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_mapping_years"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_mapping_years}{}}}
\subsection{Method \code{set_mapping_years()}}{
Set mapping years.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_mapping_years(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Integer vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_un_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_un_estimates}{}}}
\subsection{Method \code{set_un_estimates()}}{
Set UN estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_un_estimates(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_wuenic_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_wuenic_estimates}{}}}
\subsection{Method \code{set_wuenic_estimates()}}{
Set WUENIC estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_wuenic_estimates(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_national_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_national_survey}{}}}
\subsection{Method \code{set_national_survey()}}{
Set national survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_national_survey(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_regional_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_regional_survey}{}}}
\subsection{Method \code{set_regional_survey()}}{
Set regional survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_regional_survey(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_wiq_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_wiq_survey}{}}}
\subsection{Method \code{set_wiq_survey()}}{
Set WIQ survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_wiq_survey(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_area_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_area_survey}{}}}
\subsection{Method \code{set_area_survey()}}{
Set area-level survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_area_survey(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_education_survey"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_education_survey}{}}}
\subsection{Method \code{set_education_survey()}}{
Set education-level survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_education_survey(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_survey_mapping"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_survey_mapping}{}}}
\subsection{Method \code{set_survey_mapping()}}{
Set survey mapping table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_survey_mapping(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_map_mapping"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_map_mapping}{}}}
\subsection{Method \code{set_map_mapping()}}{
Set map overlay mapping.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_map_mapping(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
