% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indicator_threshold_coverage.R
\name{calculate_threshold}
\alias{calculate_threshold}
\title{Calculate Dropout Coverage for Health Indicators Below a Threshold}
\usage{
calculate_threshold(
  .data,
  admin_level = c("adminlevel_1", "district"),
  indicator = c("coverage", "dropout")
)
}
\arguments{
\item{.data}{A tibble of class \code{cd_data}.}

\item{admin_level}{The level of analysis.}

\item{indicator}{Character. The specific health indicator to evaluate. Options are:
\itemize{
\item \code{"coverage"}:coverage indicators.
\item \code{"dropout"}: dropout indicators.
}}
}
\value{
A tibble with the selected administrative level and coverage value for regions
that do not meet the below-10\% threshold for the specified indicator and year. If no regions
meet the criteria, a default row is returned with "None" and 0 as values.
}
\description{
This function filters health indicator data to identify the percentage of administrative
regions where the coverage for a specified indicator falls below a 10\% threshold for a given year.
If no regions meet the criteria (i.e., all values are below the threshold), a default output is returned.
}
\examples{
\dontrun{
  # Example usage:
  result <- calculate_threshold(data, filter_year = 2023, indicator = "zerodose", source = "dhis2")
  result
}

}
