% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_import_get_dhis2_data.R
\name{get_dhis2_hfd}
\alias{get_dhis2_hfd}
\title{Retrieve DHIS2 Health Facility Data}
\usage{
get_dhis2_hfd(country_iso3, start_date, end_date, level = 3, timeout = 3600)

get_dhis2_hfd(country_iso3, start_date, end_date, level = 3, timeout = 3600)
}
\arguments{
\item{country_iso3}{Character. ISO3 code of the country.}

\item{start_date}{Character. Start date in "YYYY-MM-DD" format.}

\item{end_date}{Character. End date in "YYYY-MM-DD" format.}

\item{timeout}{Numeric. Timeout for API calls in seconds. Default is 60.}
}
\value{
A list of class \code{cd_dhis2_hfd} containing:
\itemize{
\item \strong{\code{completeness}}: Data frame of completeness data.
\item \strong{\code{service}}: Data frame of service data.
\item \strong{\code{population}}: Data frame of population data.
\item \strong{\code{admin}}: Data frame of administrative data.
}

A list of class \code{cd_dhis2_hfd} containing:
\itemize{
\item \strong{\code{completeness}}: Data frame of completeness data.
\item \strong{\code{service}}: Data frame of service data.
\item \strong{\code{population}}: Data frame of population data.
\item \strong{\code{admin}}: Data frame of administrative data.
}
}
\description{
\code{get_dhis2_hfd} retrieves health facility data (HFD) from the DHIS2 API for a
specific country and date range. It collects completeness, service, population,
and administrative data.

\code{get_dhis2_hfd} retrieves health facility data (HFD) from the DHIS2 API for a
specific country and date range. It collects completeness, service, population,
and administrative data.
}
\examples{
\dontrun{
# Retrieve data for Kenya from January 1, 2020, to January 1, 2024
hfd_data <- get_dhis2_hfd("KEN", "2020-01-01", "2024-01-01")
}

\dontrun{
# Retrieve data for Kenya from January 1, 2020, to January 1, 2024
hfd_data <- get_dhis2_hfd("KEN", "2020-01-01", "2024-01-01")
}

}
