% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Processed Countdown 2030 Data from RDS File}
\usage{
load_data(path)
}
\arguments{
\item{path}{A string. The path to the RDS file containing the processed data.}
}
\value{
A list of class \code{cd_data} containing the merged data.
}
\description{
This function loads processed Countdown 2030 data from an RDS file.
}
\examples{
\dontrun{
  # Load previously processed data from an RDS file
  data <- load_data('processed_cd2030_data.rds')
}

}
