% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_excel_data}
\alias{load_excel_data}
\title{Load and Clean Countdown 2030 Excel Data}
\usage{
load_excel_data(
  filename,
  admin_sheet_name = "Admin_data",
  population_sheet_name = "Population_data",
  reporting_sheet_name = "Reporting_completeness",
  service_sheet_names = c("Service_data_1", "Service_data_2", "Service_data_3",
    "Vaccine_stock_data")
)
}
\arguments{
\item{filename}{A string. The path to the Excel file to be loaded.}

\item{admin_sheet_name}{A string. The name of the sheet containing administrative
data. Default is "Admin_data".}

\item{population_sheet_name}{A string. The name of the sheet containing population
data. Default is "Population_data".}

\item{reporting_sheet_name}{A string. The name of the sheet containing reporting
completeness data. Default is "Reporting_completeness".}

\item{service_sheet_names}{A vector of strings. The names of the sheets containing
service data. Default is \code{c("Service_data_1", "Service_data_2", "Service_data_3", "Vaccine_stock_data")}.}
}
\value{
A data frame of class \code{cd_data}, containing cleaned and processed data.
}
\description{
\code{load_excel_data} loads and processes data from an Excel file formatted
for Countdown 2030. It reads specified sheets, checks for duplicates, merges
data, and performs additional cleaning operations like fixing column names
and handling reporting rates.
}
\details{
This function is designed to handle Countdown 2030 data sheets, ensuring
that each sheet is properly loaded, standardized, and merged into a single data
frame. It includes robust handling for issues such as duplicate entries, missing
values in key columns, and inconsistent formatting across sheets.
}
\examples{
\dontrun{
  # Load and clean data from a Countdown 2030 Excel file
  data <- load_excel_data("countdown2030_data.xlsx")
}

}
