% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{new_countdown}
\alias{new_countdown}
\title{Create Countdown 2030 Data Object}
\usage{
new_countdown(.data, call = caller_env())
}
\arguments{
\item{.data}{A tibble. The cleaned and processed data to be converted to
\code{cd_data} class.}

\item{call}{The call environment}
}
\value{
A tibble of class \code{cd_data} with an \code{indicator_groups} attribute containing
validated and available indicator groups for analysis.
}
\description{
\code{new_countdown} creates a tibble of class \code{cd_data}, with additional checks and
validation for required indicator groups used in analysis.
}
\details{
This function adds metadata to the data for available indicator groups by
checking each group against the column names in the provided data.
The groups are categorized as follows:
\itemize{
\item \code{anc}: ANC indicators (\code{anc1}, \code{anc4}, \code{ipt2})
\item \code{idelv}: Indicators for delivery, stillbirth, and maternal deaths (\code{ideliv}, \code{instlivebirths})
\item \code{vacc}: Vaccination indicators (\code{penta1}, \code{penta3}, etc.)
}

These groups are verified for availability and attached as an attribute.
}
\examples{
\dontrun{
  # Convert processed data to `cd_data` class
  cd_data <- new_countdown(final_data)
}
}
