% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Plot Comparison between Two Indicators}
\usage{
plot_comparison(data, x_var, y_var, title, x_label, y_label)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{x_var}{A string specifying the column name for the x-axis variable.}

\item{y_var}{A string specifying the column name for the y-axis variable.}

\item{title}{The title of the plot.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}
}
\value{
A ggplot object representing the comparison plot.
}
\description{
This function creates a scatter plot comparison between two indicators for
each year, with a linear fit and R-squared value displayed.
}
\examples{
\dontrun{

  plot_comparison(cd_data, "anc1", "penta1", "Comparison of ANC1 and Penta1", "ANC1", "Penta1")
}
}
