% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save Processed Countdown 2030 Data to an RDS File}
\usage{
save_data(data, file)
}
\arguments{
\item{data}{A list of class \code{cd_data} containing the processed data to be saved.}

\item{file}{A string. The path to the RDS file where the data should be saved.}
}
\value{
No return value. This function is called for its side effects (saving
the data to an RDS file).
}
\description{
This function saves the processed Countdown 2030 data to an RDS file. It
checks that the input data is of class \code{cd_data} before saving.
}
\details{
The function only saves the \code{merged_data} portion of the \code{cd_data}
object to the specified file. If the input data does not belong to the \code{"cd_data"} class, an error is thrown.
}
\examples{
\dontrun{
  # Save processed Countdown 2030 data to an RDS file
  save_data(cd_data, "processed_cd2030_data.rds")
}

}
